src/backend/requirements.txt
# Core dependencies
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2

# NLP and text processing
nltk>=3.8.1
rouge-score>=0.1.2
scikit-learn>=1.2.2

# Async and HTTP
aiohttp>=3.8.5
httpx>=0.24.1
requests>=2.31.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=13.0.0
tqdm>=4.65.0

# Testing
pytest>=7.3.1
